Allow user to update a task
